Setup navigator w/ Form
