+++ package.json
@@ -9,18 +9,18 @@
     "test": "react-scripts test --verbose"
   },
   "dependencies": {
+    "@react-native-community/masked-view": "0.1.5",
     "@react-navigation/native": "^5.0.5",
     "@react-navigation/stack": "^5.0.5",
     "expo": "~36.0.0",
     "react": "~16.9.0",
     "react-dom": "~16.9.0",
     "react-native": "https://github.com/expo/react-native/archive/sdk-36.0.0.tar.gz",
-    "react-native-web": "~0.11.7",
-    "react-native-reanimated": "~1.4.0",
     "react-native-gesture-handler": "~1.5.0",
-    "react-native-screens": "2.0.0-alpha.12",
+    "react-native-reanimated": "~1.4.0",
     "react-native-safe-area-context": "0.6.0",
-    "@react-native-community/masked-view": "0.1.5"
+    "react-native-screens": "2.0.0-alpha.12",
+    "react-native-web": "~0.11.7"
   },
   "devDependencies": {
     "@babel/core": "^7.0.0",
